+++ src/Objects/addModel.js
@@ -15,45 +15,9 @@
 WHS.init.prototype.addModel = function(pathToModel, options) {
   'use strict';
 
-  var scope = {}; // INIT LOCAL SCOPE.
+  var scope = new api.construct(this, options, "model");
 
-  scope.root = this;
-
-  var opt = {};
-
-  scope.whsobject = true;
-  scope.releaseTime = new Date().getTime();
-
-  opt.color = options.color || 0xffffff;
-  opt.mass = options.mass || 0;
-
-  opt.pos = typeof options.pos == "object" ? options.pos : {
-    x: 0,
-    y: 0,
-    z: 0
-  };
-
-  opt.rot = typeof options.pos == "object" ? options.pos : {
-    x: 0,
-    y: 0,
-    z: 0
-  };
-
-  opt.material = options.materialOptions || {};
-  opt.geometry = options.geometryOptions || {};
-
-  scope.materialType = api.loadMaterial(opt.material)._material;
-
-  var key = 0;
-
-  WHS.objects.forEach(function(el) {
-    if (el.type == "model") {
-      key++;
-    }
-  });
-
-  scope.type = "model";
-  scope.name = opt.name || "model" + key;
+  scope.materialType = api.loadMaterial(opt.materialOptions)._material;
 
   //(new THREE.JSONLoader())
   api.JSONLoader().load(pathToModel, function(data) {
@@ -62,12 +26,7 @@ WHS.init.prototype.addModel = function(pathToModel, options) {
 
     // Visualization.
     scope.visible = new THREE.Mesh(data, scope.materialType);
-    scope.visible.position.set(opt.pos.x, opt.pos.y, opt.pos.z);
-    scope.visible.rotation.set(
-      (Math.PI / 180) * opt.rot.x,
-      (Math.PI / 180) * opt.rot.y,
-      (Math.PI / 180) * opt.rot.z
-    );
+
 
     // Physics.
     if (!options.onlyvis) {
@@ -79,11 +38,11 @@ WHS.init.prototype.addModel = function(pathToModel, options) {
 
       scope.body.linearDamping = 0.9; //default
       scope.body.addShape(scope.physic);
-      scope.body.position.set(opt.pos.x, opt.pos.y, opt.pos.z);
-      scope.body.quaternion.copy(scope.visible.quaternion);
+
       scope.body.name = scope.name;
     }
 
+    scope.build();
     scope.wrap = new api.Wrap(scope, scope.visible, scope.body);
 
   });